scripts/Scripts/db_connection.py
import psycopg2
from decouple import config

def create_db_connection():
    DATABASE_NAME = config('DATABASE_NAME')
    DATABASE_USER = config('DATABASE_USER')
    DATABASE_PASSWORD = config('DATABASE_PASSWORD')
    DATABASE_HOST = config('DATABASE_HOST')
    DATABASE_PORT = config('DATABASE_PORT')

    conn = psycopg2.connect(
        database=DATABASE_NAME,
        user=DATABASE_USER,
        password=DATABASE_PASSWORD,
        host=DATABASE_HOST,
        port=DATABASE_PORT
    )

    return conn
