scripts/Scripts/requirements.txt
python-decouple
psycopg2
matplotlib
pandas
seaborn
